Use REQUIRE_STACK where possible
We always want to check whether the stack is in the expected state.
Also, NEVER use REQUIRE_STACK with a temporary that changes the stack:
the value goes into a decltype() expression, which means that it's
evaluated and won't change the actual Lua stack. Also, C++17 doesn't
allow lambdas in those decltype() expressions.